Exceptions

+++ src/main/java/com/capstone/donorhub/service/AdminServiceImpl.java
@@ -1,5 +1,6 @@
 package com.capstone.donorhub.service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,57 +23,109 @@ public class AdminServiceImpl {
 	@Autowired
 	private PasswordEncoder passwordEncoder;
 
-	//GetAllUsers
+	
+	//---------------------------------------------
 	public List<User> getAllUsers() {
 		return userRepository.findAll();
 	}
 
-	//Add a new User
+	//---------------------------------------------
 	public User saveUser(User user) {
 		user.setPassword(this.passwordEncoder.encode(user.getPassword()));
 		return userRepository.save(user);
 	}
 
-	//GetAllItems
+	//-----------------------------------------------
 	public List<Items> getAllItem() {
 		return itemRepository.findAll();
 	}
 
-	//GetASignleUser
+	//------------------------------------------------
+	
 	public User getSingleUser(int id) {
+		
+		User c=null;
 		java.util.Optional<User> userOptional = userRepository.findById(id);
 		if (userOptional.isPresent()) {
 			return userOptional.get();
 		}
-		throw new RuntimeException("User not found for id: " + id);
+		else
+		{
+			return c;
+		}
+//		throw new RuntimeException("User not found for id: " + id);
 
 	}
+	
+	//----------------------------------------------------------
 
-	//DeleteAUser
-	public void deleteUser(int id) {
+	public String deleteUser(int id) {
+		java.util.Optional<User> userOptional = userRepository.findById(id);
+		
+		if (userOptional.isPresent()) {
 		userRepository.deleteById(id);
+		return "user deleted";
+		}
+		else
+		{
+			return "no such user is present";
+		}
+		
 
 	}
+	
+	//-------------------------------------------------------
 
-	//UpdateUser
 	public User updateUser(User user) {
 		user.setPassword(this.passwordEncoder.encode(user.getPassword()));
 		return userRepository.save(user);
 	}
 
-	//DeleteAnItem
-	public void deleteItem(int id) {
+	//------------------------------------------------------
+	public String deleteItem(int id) {
+		
+		java.util.Optional<Items> itemOptional = itemRepository.findById(id);
+		if (itemOptional.isPresent()) {
 		itemRepository.deleteById(id);
-
-	}
-
-	//GetUserByName
-	public List<User> getUserByName(String name) {
-
-		return userRepository.findByName(name);
+		return "item deleted";
+		}
+		
+		else
+		{
+			return "No such item is present";
+		}
 	}
 	
-	//UpdateAccountStatus
+	//----------------------------------------------------------
+
+	public User updatUser(User user) {
+		return userRepository.save(user);
+	}
+	
+	//----------------------------------------------------------
+
+	public List<User> getUserByName(String name) {
+		
+		List<User> dummy = new ArrayList<User>();
+		
+		List<User> user = userRepository.findByName(name);
+		
+		if(user.isEmpty())
+		{
+			return dummy;
+		}
+		else
+		{
+			return userRepository.findByName(name);
+
+		}
+
+	}
+	
+	//-----------------------------------------------------------
+	
+	
+	
 	 public User updateUserAccountStatus(int userId, String accountStatus) {
 	        Optional<User> optionalUser = userRepository.findById(userId);
 	        if (optionalUser.isPresent()) {
@@ -83,6 +136,7 @@ public class AdminServiceImpl {
 	        } else {
 	            throw new RuntimeException("User not found with ID: " + userId);
 	        }
-	        
+
 }
+	 //--------------------------------------------------------------------
 }
